src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/alternate-stylesheets-expected.txt

CONSOLE ERROR: line 5: Uncaught TypeError: host.createShadowRoot is not a function
Title attribute should be ignored for style elements in shadow trees.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL getComputedStyle(root.querySelector('div')).color should be rgb(0, 128, 0). Threw exception TypeError: Cannot read property 'querySelector' of undefined
PASS getComputedStyle(alternate).color is "rgb(0, 128, 0)"
FAIL host.shadowRoot.styleSheets[0].title should be null. Threw exception TypeError: Cannot read property 'styleSheets' of null
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE

Should be green
